converting seconds to string is remaining
Berkeley’s Algorithm is a clock synchronisation technique used in distributed systems. The algorithm assumes that each machine node in the network either doesn’t have an accurate time source or doesn’t possess an UTC server. 

The masterTime is initialised to a default value in program. The slave time is taken as input in hh:mm:ss 24 hour format. The output is synchronised time in seconds in terms of number of seconds lapsed since midnight.